Move around changeFilename tests
