+++ test/renderer/reducers/document_spec.js
@@ -317,35 +317,3 @@ describe('overwriteMetadata', () => {
     expect(state.document.getIn(['notebook', 'metadata', 'name'])).to.equal("javascript");
   });
 });
-
-describe('changeFilename', () => {
-  it('returns the same originalState if filename is undefined', () => {
-    const originalState = {
-      document: new DocumentRecord({
-        filename: 'original',
-      })
-    };
-
-    const action = {
-      type: constants.CHANGE_FILENAME,
-    };
-
-    const state = reducers(originalState, action);
-    expect(state.document.filename).to.equal('original');
-  });
-  it('sets the filename if given a valid one', () => {
-    const originalState = {
-      document: new DocumentRecord({
-        filename: 'original',
-      })
-    };
-
-    const action = {
-      type: constants.CHANGE_FILENAME,
-      filename: 'test.ipynb',
-    };
-
-    const state = reducers(originalState, action);
-    expect(state.document.filename).to.equal('test.ipynb');
-  });
-});
